change return type of buildHeader() from size_t to uint8_t as size_t it not needed here

+++ src/PubSubClient.h
@@ -151,11 +151,7 @@ class PubSubClient : public Print {
     bool readByte(uint8_t* result, uint16_t* index);
     bool write(uint8_t header, uint8_t* buf, size_t length);
     size_t writeString(const char* string, uint8_t* buf, size_t pos);
-    // Build up the header ready to send
-    // Returns the size of the header
-    // Note: the header is built at the end of the first MQTT_MAX_HEADER_SIZE bytes, so will start
-    //       (MQTT_MAX_HEADER_SIZE - <returned size>) bytes into the buffer
-    size_t buildHeader(uint8_t header, uint8_t* buf, size_t length);
+    uint8_t buildHeader(uint8_t header, uint8_t* buf, size_t length);
     IPAddress ip;
     char* domain;
     uint16_t port;
